requirements.txt
#pyaudio==0.2.13
openai==1.6.1
python-dotenv==1.0.0
blockfrost-python==0.6.0
click==8.1.7
pydub==0.25.1
SpeechRecognition==3.10.0
pyttsx3==2.90
python-telegram-bot==13.15 